@@ -7,8 +7,7 @@
     "graphql": "15.8.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "ts-invariant": "0.10.3",
-    "web-vitals": "2.1.4"
+    "ts-invariant": "0.10.3"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "5.16.4",
@@ -21,7 +20,7 @@
     "eslint-config-checkout": "workspace:*",
     "react-scripts": "5.0.1",
     "tsconfig": "workspace:*",
-    "typescript": "4.7.4"
+    "typescript": "4.8.4"
   },
   "scripts": {
     "dev": "react-scripts start",
