@@ -21,51 +21,51 @@
     "graphql": "15.8.0"
   },
   "dependencies": {
-    "@adyen/adyen-web": "5.26.0",
+    "@adyen/adyen-web": "5.28.2",
     "@adyen/api-library": "12.0.0",
     "@material-ui/core": "^4.12.3",
     "@material-ui/icons": "^4.11.3",
     "@material-ui/lab": "^4.0.0-alpha.60",
     "@material-ui/styles": "^4.11.5",
     "@mollie/api-client": "^3.6.0",
-    "@saleor/app-sdk": "~0.9.1",
+    "@saleor/app-sdk": "~0.17.0",
     "@saleor/checkout-storefront": "workspace:*",
-    "@saleor/macaw-ui": "^0.4.0",
+    "@saleor/macaw-ui": "^0.6.6",
     "@sentry/nextjs": "7.16.0",
-    "@urql/exchange-auth": "^0.1.7",
-    "@urql/exchange-multipart-fetch": "^0.1.14",
+    "@urql/exchange-auth": "^1.0.0",
+    "@urql/exchange-multipart-fetch": "^1.0.1",
     "checkout-common": "workspace:*",
     "clsx": "^1.1.1",
     "content-type": "^1.0.4",
     "crypto-js": "^4.1.1",
     "currency.js": "^2.0.4",
-    "graphql": "15.8.0",
+    "graphql": "16.6.0",
     "graphql-tag": "^2.12.6",
-    "jose": "^4.8.1",
+    "jose": "^4.10.3",
     "jsonwebtoken": "^8.5.1",
-    "jwks-rsa": "^2.1.2",
+    "jwks-rsa": "^2.1.5",
     "next": "12.3.1",
     "raw-body": "^2.5.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-hook-form": "^7.27.1",
-    "react-intl": "^6.1.1",
-    "react-is": "^17.0.2",
-    "retes": "^0.33.0",
-    "stripe": "10.2.0",
+    "react-hook-form": "^7.38.0",
+    "react-intl": "^6.2.1",
+    "react-is": "^18.2.0",
+    "retes": "^0.34.0",
+    "stripe": "10.15.0",
     "ts-invariant": "0.10.3",
     "url-join": "4.0.1",
-    "urql": "^2.2.0",
+    "urql": "^3.0.3",
     "yup": "1.0.0-beta.7"
   },
   "devDependencies": {
-    "@formatjs/cli": "^5.1.0",
+    "@formatjs/cli": "^5.1.3",
     "@graphql-codegen/add": "^3.2.0",
-    "@graphql-codegen/cli": "^2.8.1",
+    "@graphql-codegen/cli": "^2.13.7",
     "@graphql-codegen/introspection": "^2.2.0",
-    "@graphql-codegen/typescript": "^2.7.1",
-    "@graphql-codegen/typescript-operations": "^2.5.1",
-    "@graphql-codegen/typescript-urql": "^3.6.1",
+    "@graphql-codegen/typescript": "^2.8.0",
+    "@graphql-codegen/typescript-operations": "^2.5.5",
+    "@graphql-codegen/typescript-urql": "^3.7.2",
     "@graphql-codegen/urql-introspection": "^2.2.0",
     "@pollyjs/adapter-fetch": "^6.0.5",
     "@pollyjs/adapter-node-http": "^6.0.5",
@@ -78,28 +78,28 @@
     "@types/lodash-es": "^4.17.6",
     "@types/node": "^16.11.22",
     "@types/omit-deep-lodash": "^1.1.1",
-    "@types/react": "^18.0.15",
-    "@types/react-dom": "^18.0.6",
+    "@types/react": "^18.0.22",
+    "@types/react-dom": "^18.0.7",
     "@types/setup-polly-jest": "^0.5.1",
     "@types/url-join": "4.0.1",
     "dotenv": "^16.0.1",
     "env-vars": "workspace:*",
-    "eslint": "^8.22.0",
+    "eslint": "^8.26.0",
     "eslint-config-checkout": "workspace:*",
-    "eslint-plugin-formatjs": "^4.2.2",
-    "fast-check": "3.1.1",
-    "headers-polyfill": "^3.0.9",
+    "eslint-plugin-formatjs": "^4.3.4",
+    "fast-check": "3.3.0",
+    "headers-polyfill": "^3.1.2",
     "jest": "^27.5.1",
     "json-easy-strip": "^1.0.6",
-    "msw": "^0.43.1",
+    "msw": "^0.47.4",
     "next-test-api-route-handler": "^3.1.7",
-    "next-transpile-modules": "9.0.0",
+    "next-transpile-modules": "9.1.0",
     "node-mocks-http": "^1.11.0",
     "omit-deep-lodash": "^1.1.7",
     "setup-polly-jest": "^0.11.0",
     "ts-jest": "^27.1.4",
     "tsconfig": "workspace:*",
-    "typescript": "4.7.4"
+    "typescript": "4.8.4"
   },
   "engines": {
     "npm": ">=7.0.0",
