@@ -18,7 +18,7 @@
     "clean": "rm -rf .next .turbo"
   },
   "dependencies": {
-    "@apollo/client": "3.5.10",
+    "@apollo/client": "3.7.0",
     "@formatjs/swc-plugin": "^1.4.0",
     "@graphql-codegen/typescript-apollo-client-helpers": "^2.1.15",
     "@headlessui/react": "^1.6.0",
@@ -73,7 +73,7 @@
     "react-hook-form": "^7.29.0",
     "react-use": "^17.3.2",
     "tailwindcss": "3.0.23",
-    "typescript": "4.7.4"
+    "typescript": "4.8.4"
   },
   "engines": {
     "npm": ">=7.0.0",
