@@ -32,13 +32,13 @@
     "cy:open": "cypress open"
   },
   "devDependencies": {
-    "cypress": "^10.1.0",
+    "cypress": "^10.10.0",
     "env-cmd": "^10.1.0",
     "husky": "^8.0.1",
     "lint-staged": "13.0.3",
     "prettier": "^2.7.1",
-    "turbo": "^1.4.6",
-    "typescript": "4.7.4"
+    "turbo": "^1.6.1",
+    "typescript": "4.8.4"
   },
   "engines": {
     "npm": ">=7.0.0",
