@@ -14,26 +14,29 @@
     "lint:staged": "eslint --fix --quiet"
   },
   "dependencies": {
-    "@adyen/adyen-web": "5.26.0",
+    "@adyen/adyen-web": "5.28.2",
     "@adyen/api-library": "12.0.0",
     "@material-ui/core": "^4.12.3",
+    "autoprefixer": "10.4.12",
+    "postcss-import": "15.0.0",
     "rollup-plugin-esbuild": "^4.9.1",
+    "tailwindcss": "3.2.1",
     "yup": "1.0.0-beta.7"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^22.0.1",
-    "@rollup/plugin-image": "2.1.1",
-    "@rollup/plugin-json": "4.1.0",
-    "@rollup/plugin-node-resolve": "^13.3.0",
-    "@rollup/plugin-typescript": "^8.3.3",
+    "@rollup/plugin-commonjs": "^23.0.2",
+    "@rollup/plugin-image": "3.0.1",
+    "@rollup/plugin-json": "5.0.1",
+    "@rollup/plugin-node-resolve": "^15.0.1",
+    "@rollup/plugin-typescript": "^9.0.2",
     "eslint-config-checkout": "workspace:*",
-    "rollup": "^2.70.2",
-    "rollup-plugin-dts": "^4.2.2",
+    "rollup": "^3.2.3",
+    "rollup-plugin-dts": "^5.0.0",
     "rollup-plugin-peer-deps-external": "^2.2.4",
     "rollup-plugin-postcss": "^4.0.2",
     "rollup-plugin-terser": "^7.0.2",
     "tsconfig": "workspace:*",
-    "typescript": "4.7.4"
+    "typescript": "4.8.4"
   },
   "engines": {
     "npm": ">=7.0.0",
