@@ -31,71 +31,71 @@
     "@types/react": "^18.0.15"
   },
   "dependencies": {
-    "@adyen/adyen-web": "5.26.0",
+    "@adyen/adyen-web": "5.28.2",
     "@adyen/api-library": "12.0.0",
-    "@apollo/client": "^3.5.10",
+    "@apollo/client": "3.7.0",
     "@headlessui/react": "^1.6.0",
     "@hookform/error-message": "^2.0.0",
     "@react-types/button": "^3.4.1",
     "@saleor/sdk": "^0.4.3",
     "@saleor/ui-kit": "workspace:*",
-    "@stripe/react-stripe-js": "^1.7.0",
-    "@stripe/stripe-js": "^1.26.0",
+    "@stripe/react-stripe-js": "^1.13.0",
+    "@stripe/stripe-js": "^1.42.0",
     "checkout-common": "workspace:*",
     "clsx": "^1.1.1",
-    "concurrently": "7.3.0",
+    "concurrently": "7.5.0",
     "hex-rgb": "^5.0.0",
     "lodash-es": "^4.17.21",
     "query-string": "^7.1.1",
     "react": "^18.2.0",
     "react-error-boundary": "^3.1.4",
-    "react-hook-form": "^7.34.2",
-    "react-intl": "^6.1.1",
+    "react-hook-form": "^7.38.0",
+    "react-intl": "^6.2.1",
     "react-test-renderer": "^18.2.0",
     "react-toastify": "^9.0.4",
-    "tailwindcss": "^3.0.18",
+    "tailwindcss": "^3.2.1",
     "ts-invariant": "0.10.3",
     "url-join": "4.0.1",
-    "urql": "^2.2.3",
+    "urql": "^3.0.3",
     "yup": "1.0.0-beta.7"
   },
   "devDependencies": {
-    "@formatjs/cli": "5.1.0",
+    "@formatjs/cli": "5.1.3",
     "@graphql-codegen/add": "^3.2.0",
-    "@graphql-codegen/cli": "2.8.1",
-    "@graphql-codegen/introspection": "2.2.0",
-    "@graphql-codegen/typescript": "2.7.1",
-    "@graphql-codegen/typescript-operations": "2.5.1",
-    "@graphql-codegen/typescript-urql": "^3.6.1",
-    "@graphql-codegen/urql-introspection": "2.2.0",
-    "@rollup/plugin-commonjs": "^23.0.0",
-    "@rollup/plugin-image": "3.0.0",
-    "@rollup/plugin-json": "5.0.0",
-    "@rollup/plugin-node-resolve": "^15.0.0",
-    "@rollup/plugin-typescript": "^9.0.1",
+    "@graphql-codegen/cli": "2.13.7",
+    "@graphql-codegen/introspection": "2.2.1",
+    "@graphql-codegen/typescript": "2.8.0",
+    "@graphql-codegen/typescript-operations": "2.5.5",
+    "@graphql-codegen/typescript-urql": "^3.7.2",
+    "@graphql-codegen/urql-introspection": "2.2.1",
+    "@rollup/plugin-commonjs": "^23.0.2",
+    "@rollup/plugin-image": "3.0.1",
+    "@rollup/plugin-json": "5.0.1",
+    "@rollup/plugin-node-resolve": "^15.0.1",
+    "@rollup/plugin-typescript": "^9.0.2",
     "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/react-hooks": "^8.0.1",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/user-event": "^14.4.3",
     "@types/jest": "^27.5.2",
     "@types/lodash-es": "^4.17.6",
     "@types/node": "^16.11.22",
-    "@types/react": "^18.0.15",
-    "@types/react-dom": "^18.0.6",
+    "@types/react": "^18.0.22",
+    "@types/react-dom": "^18.0.7",
     "@types/url-join": "4.0.1",
     "@types/webpack-env": "^1.16.3",
-    "autoprefixer": "9.8.6",
+    "autoprefixer": "10.4.12",
     "env-vars": "workspace:*",
-    "eslint": "^8.22.0",
+    "eslint": "^8.26.0",
     "eslint-config-checkout": "workspace:*",
-    "eslint-plugin-formatjs": "^4.2.2",
+    "eslint-plugin-formatjs": "^4.3.4",
     "graphql-tag": "^2.12.6",
     "jest": "^27.5.1",
     "next": "12.3.1",
-    "postcss": "^8.4.6",
-    "postcss-import": "14.1.0",
+    "postcss": "^8.4.18",
+    "postcss-import": "15.0.0",
     "prettier": "^2.5.1",
-    "rollup": "^3.2.2",
+    "rollup": "^3.2.3",
     "rollup-plugin-dts": "^5.0.0",
     "rollup-plugin-peer-deps-external": "^2.2.4",
     "rollup-plugin-postcss": "^4.0.2",
@@ -104,7 +104,6 @@
     "tsconfig": "workspace:*",
     "tslib": "^2.4.0",
     "typescript": "4.8.4",
-    "web-vitals": "^2.1.4",
     "wonka": "^6.0.0"
   },
   "engines": {
