@@ -20,7 +20,7 @@
     "lint": "eslint",
     "lint:staged": "eslint --fix --quiet",
     "check-types": "tsc",
-    "clean": "rm -rf .next .turbo",
+    "clean": "rm -rf .next .turbo dist",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
     "release": "release-it --npm.pnpm"
@@ -34,44 +34,46 @@
     "react-dom": "^18.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.17.8",
-    "@rollup/plugin-commonjs": "^22.0.0",
-    "@rollup/plugin-node-resolve": "^13.2.1",
-    "@rollup/plugin-typescript": "^8.3.2",
-    "@storybook/addon-actions": "^6.4.22",
-    "@storybook/addon-essentials": "^6.4.22",
-    "@storybook/addon-interactions": "^6.4.22",
-    "@storybook/addon-links": "^6.4.22",
+    "@babel/core": "^7.19.6",
+    "@rollup/plugin-commonjs": "^23.0.2",
+    "@rollup/plugin-node-resolve": "^15.0.1",
+    "@rollup/plugin-typescript": "^9.0.2",
+    "@storybook/addon-actions": "^6.5.13",
+    "@storybook/addon-essentials": "^6.5.13",
+    "@storybook/addon-interactions": "^6.5.13",
+    "@storybook/addon-links": "^6.5.13",
     "@storybook/addon-postcss": "^2.0.0",
-    "@storybook/react": "v6.5.0-beta.7",
-    "@storybook/testing-library": "^0.0.11",
+    "@storybook/builder-webpack5": "6.5.13",
+    "@storybook/manager-webpack5": "6.5.13",
+    "@storybook/react": "6.5.13",
+    "@storybook/testing-library": "^0.0.13",
     "@types/node": "^16.11.22",
-    "@types/react": "^18.0.15",
-    "@types/react-dom": "^18.0.6",
+    "@types/react": "^18.0.22",
+    "@types/react-dom": "^18.0.7",
     "autoprefixer": "^10.4.4",
     "babel-loader": "^8.2.4",
-    "css-loader": "^3.6.0",
+    "css-loader": "^6.7.1",
     "css-modules-typescript-loader": "^4.0.1",
     "eslint-config-checkout": "workspace:*",
-    "postcss": "^8.4.12",
-    "postcss-loader": "4.3.0",
+    "postcss": "^8.4.18",
+    "postcss-loader": "7.0.1",
     "react": "^18.2.0",
     "react-docgen-typescript": "^2.2.2",
     "react-dom": "^18.2.0",
     "release-it": "^15.0.0",
-    "rollup-plugin-dts": "^4.2.1",
+    "rollup-plugin-dts": "^5.0.0",
     "rollup-plugin-peer-deps-external": "^2.2.4",
     "rollup-plugin-postcss": "^4.0.2",
     "rollup-plugin-terser": "^7.0.2",
-    "style-loader": "^1.3.0",
-    "tailwindcss": "3.0.18",
+    "style-loader": "^3.3.1",
+    "tailwindcss": "3.2.1",
     "tsconfig": "workspace:*",
-    "typescript": "4.7.4"
+    "typescript": "4.8.4"
   },
   "dependencies": {
-    "@headlessui/react": "1.6.0",
+    "@headlessui/react": "1.7.3",
     "clsx": "^1.1.1",
-    "rollup": "^2.70.2",
+    "rollup": "^3.2.3",
     "tslib": "^2.4.0"
   },
   "engines": {
