@@ -1308,7 +1308,7 @@ export type AttributeDeleted = Event & {
 };
 
 /** An enumeration. */
-export type AttributeEntityTypeEnum = "PAGE" | "PRODUCT" | "PRODUCT_VARIANT";
+export type AttributeEntityTypeEnum = "PAGE" | "PRODUCT";
 
 export type AttributeError = {
   __typename?: "AttributeError";
@@ -1345,7 +1345,6 @@ export type AttributeFilterInput = {
   isVariantOnly?: InputMaybe<Scalars["Boolean"]>;
   metadata?: InputMaybe<Array<MetadataFilter>>;
   search?: InputMaybe<Scalars["String"]>;
-  slugs?: InputMaybe<Array<Scalars["String"]>>;
   type?: InputMaybe<AttributeTypeEnum>;
   valueRequired?: InputMaybe<Scalars["Boolean"]>;
   visibleInStorefront?: InputMaybe<Scalars["Boolean"]>;
@@ -1615,18 +1614,12 @@ export type AttributeValueCreateInput = {
   fileUrl?: InputMaybe<Scalars["String"]>;
   /** Name of a value displayed in the interface. */
   name: Scalars["String"];
-  /**
-   * Represents the text of the attribute value, plain text without formating.
-   *
-   * DEPRECATED: this field will be removed in Saleor 4.0.The plain text attribute hasn't got predefined value, so can be specified only from instance that supports the given attribute.
-   */
+  /** Represents the text of the attribute value, plain text without formating. */
   plainText?: InputMaybe<Scalars["String"]>;
   /**
    * Represents the text of the attribute value, includes formatting.
    *
    * Rich text format. For reference see https://editorjs.io/
-   *
-   * DEPRECATED: this field will be removed in Saleor 4.0.The rich text attribute hasn't got predefined value, so can be specified only from instance that supports the given attribute.
    */
   richText?: InputMaybe<Scalars["JSONString"]>;
   /** Represent value of the attribute value (e.g. color values for swatch attributes). */
@@ -1806,18 +1799,12 @@ export type AttributeValueUpdateInput = {
   fileUrl?: InputMaybe<Scalars["String"]>;
   /** Name of a value displayed in the interface. */
   name?: InputMaybe<Scalars["String"]>;
-  /**
-   * Represents the text of the attribute value, plain text without formating.
-   *
-   * DEPRECATED: this field will be removed in Saleor 4.0.The plain text attribute hasn't got predefined value, so can be specified only from instance that supports the given attribute.
-   */
+  /** Represents the text of the attribute value, plain text without formating. */
   plainText?: InputMaybe<Scalars["String"]>;
   /**
    * Represents the text of the attribute value, includes formatting.
    *
    * Rich text format. For reference see https://editorjs.io/
-   *
-   * DEPRECATED: this field will be removed in Saleor 4.0.The rich text attribute hasn't got predefined value, so can be specified only from instance that supports the given attribute.
    */
   richText?: InputMaybe<Scalars["JSONString"]>;
   /** Represent value of the attribute value (e.g. color values for swatch attributes). */
@@ -1888,26 +1875,6 @@ export type BulkStockError = {
   message?: Maybe<Scalars["String"]>;
   /** List of attribute values IDs which causes the error. */
   values?: Maybe<Array<Scalars["ID"]>>;
-};
-
-/**
- * Synchronous webhook for calculating checkout/order taxes.
- *
- * Added in Saleor 3.7.
- *
- * Note: this API is currently in Feature Preview and can be subject to changes at later point.
- */
-export type CalculateTaxes = Event & {
-  __typename?: "CalculateTaxes";
-  /** Time of the event. */
-  issuedAt?: Maybe<Scalars["DateTime"]>;
-  /** The user or application that triggered the event. */
-  issuingPrincipal?: Maybe<IssuingPrincipal>;
-  /** The application receiving the webhook. */
-  recipient?: Maybe<App>;
-  taxBase: TaxableObject;
-  /** Saleor version that triggered the event. */
-  version?: Maybe<Scalars["String"]>;
 };
 
 export type CardInput = {
@@ -2168,7 +2135,6 @@ export type CategoryFilterInput = {
   ids?: InputMaybe<Array<Scalars["ID"]>>;
   metadata?: InputMaybe<Array<MetadataFilter>>;
   search?: InputMaybe<Scalars["String"]>;
-  slugs?: InputMaybe<Array<Scalars["String"]>>;
 };
 
 export type CategoryInput = {
@@ -2777,14 +2743,6 @@ export type Checkout = Node &
     stockReservationExpires?: Maybe<Scalars["DateTime"]>;
     /** The price of the checkout before shipping, with taxes included. */
     subtotalPrice: TaxedMoney;
-    /**
-     * Returns True if checkout has to be exempt from taxes.
-     *
-     * Added in Saleor 3.8.
-     *
-     * Note: this API is currently in Feature Preview and can be subject to changes at later point.
-     */
-    taxExemption: Scalars["Boolean"];
     /** The checkout's token. */
     token: Scalars["UUID"];
     /** The sum of the the checkout line prices, with all the taxes,shipping costs, and discounts included. */
@@ -3200,12 +3158,6 @@ export type CheckoutLineInput = {
    * Note: this API is currently in Feature Preview and can be subject to changes at later point.
    */
   forceNewLine?: InputMaybe<Scalars["Boolean"]>;
-  /**
-   * Fields required to update the object's metadata.
-   *
-   * Added in Saleor 3.8.
-   */
-  metadata?: InputMaybe<Array<MetadataInput>>;
   /**
    * Custom price of the item. Can be set only by apps with `HANDLE_CHECKOUTS` permission. When the line with the same variant will be provided multiple times, the last price will be used.
    *
@@ -3732,7 +3684,6 @@ export type CollectionFilterInput = {
   metadata?: InputMaybe<Array<MetadataFilter>>;
   published?: InputMaybe<CollectionPublished>;
   search?: InputMaybe<Scalars["String"]>;
-  slugs?: InputMaybe<Array<Scalars["String"]>>;
 };
 
 export type CollectionInput = {
@@ -4399,12 +4350,6 @@ export type CustomerEventsEnum =
 
 export type CustomerFilterInput = {
   dateJoined?: InputMaybe<DateRangeInput>;
-  /**
-   * Filter by ids.
-   *
-   * Added in Saleor 3.8.
-   */
-  ids?: InputMaybe<Array<Scalars["ID"]>>;
   metadata?: InputMaybe<Array<MetadataFilter>>;
   numberOfOrders?: InputMaybe<IntRangeInput>;
   placedOrders?: InputMaybe<DateRangeInput>;
@@ -7850,7 +7795,6 @@ export type MenuFilterInput = {
   metadata?: InputMaybe<Array<MetadataFilter>>;
   search?: InputMaybe<Scalars["String"]>;
   slug?: InputMaybe<Array<Scalars["String"]>>;
-  slugs?: InputMaybe<Array<Scalars["String"]>>;
 };
 
 export type MenuInput = {
@@ -9780,16 +9724,6 @@ export type Mutation = {
    * Requires one of the following permissions: MANAGE_STAFF.
    */
   staffUpdate?: Maybe<StaffUpdate>;
-  /**
-   * Exempt checkout or order from charging the taxes. When tax exemption is enabled, taxes won't be charged for the checkout or order. Taxes may still be calculated in cases when product prices are entered with the tax included and the net price needs to be known.
-   *
-   * Added in Saleor 3.8.
-   *
-   * Note: this API is currently in Feature Preview and can be subject to changes at later point.
-   *
-   * Requires one of the following permissions: MANAGE_TAXES.
-   */
-  taxExemptionManage?: Maybe<TaxExemptionManage>;
   /** Create JWT token. */
   tokenCreate?: Maybe<CreateToken>;
   /** Refresh JWT token. Mutation tries to take refreshToken from the input.If it fails it will try to take refreshToken from the http-only cookie -refreshToken. csrfToken is required when refreshToken is provided as a cookie. */
@@ -10179,7 +10113,6 @@ export type MutationCheckoutBillingAddressUpdateArgs = {
 export type MutationCheckoutCompleteArgs = {
   checkoutId?: InputMaybe<Scalars["ID"]>;
   id?: InputMaybe<Scalars["ID"]>;
-  metadata?: InputMaybe<Array<MetadataInput>>;
   paymentData?: InputMaybe<Scalars["JSONString"]>;
   redirectUrl?: InputMaybe<Scalars["String"]>;
   storeSource?: InputMaybe<Scalars["Boolean"]>;
@@ -10604,8 +10537,6 @@ export type MutationOrderConfirmArgs = {
 
 export type MutationOrderCreateFromCheckoutArgs = {
   id: Scalars["ID"];
-  metadata?: InputMaybe<Array<MetadataInput>>;
-  privateMetadata?: InputMaybe<Array<MetadataInput>>;
   removeCheckout?: InputMaybe<Scalars["Boolean"]>;
 };
 
@@ -11140,11 +11071,6 @@ export type MutationStaffUpdateArgs = {
   input: StaffUpdateInput;
 };
 
-export type MutationTaxExemptionManageArgs = {
-  id: Scalars["ID"];
-  taxExemption: Scalars["Boolean"];
-};
-
 export type MutationTokenCreateArgs = {
   email: Scalars["String"];
   password: Scalars["String"];
@@ -11386,7 +11312,7 @@ export type Order = Node &
     created: Scalars["DateTime"];
     customerNote: Scalars["String"];
     /**
-     * The delivery method selected for this order.
+     * The delivery method selected for this checkout.
      *
      * Added in Saleor 3.1.
      *
@@ -11502,14 +11428,6 @@ export type Order = Node &
     statusDisplay: Scalars["String"];
     /** The sum of line prices not including shipping. */
     subtotal: TaxedMoney;
-    /**
-     * Returns True if order has to be exempt from taxes.
-     *
-     * Added in Saleor 3.8.
-     *
-     * Note: this API is currently in Feature Preview and can be subject to changes at later point.
-     */
-    taxExemption: Scalars["Boolean"];
     /** @deprecated This field will be removed in Saleor 4.0. Use `id` instead. */
     token: Scalars["String"];
     /** Total amount of the order. */
@@ -13026,7 +12944,6 @@ export type PageFilterInput = {
   metadata?: InputMaybe<Array<MetadataFilter>>;
   pageTypes?: InputMaybe<Array<Scalars["ID"]>>;
   search?: InputMaybe<Scalars["String"]>;
-  slugs?: InputMaybe<Array<Scalars["String"]>>;
 };
 
 /** The Relay compliant `PageInfo` type, containing data necessary to paginate this connection. */
@@ -13400,7 +13317,6 @@ export type PageTypeDeleted = Event & {
 
 export type PageTypeFilterInput = {
   search?: InputMaybe<Scalars["String"]>;
-  slugs?: InputMaybe<Array<Scalars["String"]>>;
 };
 
 /**
@@ -13796,12 +13712,6 @@ export type PaymentErrorCode =
 
 export type PaymentFilterInput = {
   checkouts?: InputMaybe<Array<Scalars["ID"]>>;
-  /**
-   * Filter by ids.
-   *
-   * Added in Saleor 3.8.
-   */
-  ids?: InputMaybe<Array<Scalars["ID"]>>;
 };
 
 /** Available payment gateway backend with configuration necessary to setup client. */
@@ -14022,7 +13932,6 @@ export type PermissionEnum =
   | "MANAGE_SETTINGS"
   | "MANAGE_SHIPPING"
   | "MANAGE_STAFF"
-  | "MANAGE_TAXES"
   | "MANAGE_TRANSLATIONS"
   | "MANAGE_USERS";
 
@@ -14940,7 +14849,6 @@ export type ProductFilterInput = {
   price?: InputMaybe<PriceRangeInput>;
   productTypes?: InputMaybe<Array<Scalars["ID"]>>;
   search?: InputMaybe<Scalars["String"]>;
-  slugs?: InputMaybe<Array<Scalars["String"]>>;
   /** Filter by variants having specific stock status. */
   stockAvailability?: InputMaybe<StockAvailability>;
   stocks?: InputMaybe<ProductStockFilterInput>;
@@ -15474,7 +15382,6 @@ export type ProductTypeFilterInput = {
   metadata?: InputMaybe<Array<MetadataFilter>>;
   productType?: InputMaybe<ProductTypeEnum>;
   search?: InputMaybe<Scalars["String"]>;
-  slugs?: InputMaybe<Array<Scalars["String"]>>;
 };
 
 export type ProductTypeInput = {
@@ -19424,42 +19331,6 @@ export type Subscription = {
   event?: Maybe<Event>;
 };
 
-/**
- * Exempt checkout or order from charging the taxes. When tax exemption is enabled, taxes won't be charged for the checkout or order. Taxes may still be calculated in cases when product prices are entered with the tax included and the net price needs to be known.
- *
- * Added in Saleor 3.8.
- *
- * Note: this API is currently in Feature Preview and can be subject to changes at later point.
- *
- * Requires one of the following permissions: MANAGE_TAXES.
- */
-export type TaxExemptionManage = {
-  __typename?: "TaxExemptionManage";
-  errors: Array<TaxExemptionManageError>;
-  taxableObject?: Maybe<TaxSourceObject>;
-};
-
-export type TaxExemptionManageError = {
-  __typename?: "TaxExemptionManageError";
-  /** The error code. */
-  code: TaxExemptionManageErrorCode;
-  /** Name of a field that caused the error. A value of `null` indicates that the error isn't associated with a particular field. */
-  field?: Maybe<Scalars["String"]>;
-  /** The error message. */
-  message?: Maybe<Scalars["String"]>;
-};
-
-/** An enumeration. */
-export type TaxExemptionManageErrorCode =
-  | "GRAPHQL_ERROR"
-  | "INVALID"
-  | "NOT_EDITABLE_ORDER"
-  | "NOT_FOUND";
-
-export type TaxSourceLine = CheckoutLine | OrderLine;
-
-export type TaxSourceObject = Checkout | Order;
-
 /** Representation of tax types fetched from tax gateway. */
 export type TaxType = {
   __typename?: "TaxType";
@@ -19467,55 +19338,6 @@ export type TaxType = {
   description?: Maybe<Scalars["String"]>;
   /** External tax code used to identify given tax group. */
   taxCode?: Maybe<Scalars["String"]>;
-};
-
-/** Taxable object. */
-export type TaxableObject = {
-  __typename?: "TaxableObject";
-  /** The address data. */
-  address?: Maybe<Address>;
-  channel: Channel;
-  /** The currency of the object. */
-  currency: Scalars["String"];
-  /** List of discounts. */
-  discounts: Array<TaxableObjectDiscount>;
-  /** List of lines assigned to the object. */
-  lines: Array<TaxableObjectLine>;
-  /** Determines if prices contain entered tax.. */
-  pricesEnteredWithTax: Scalars["Boolean"];
-  /** The price of shipping method. */
-  shippingPrice: Money;
-  /** The source object related to this tax object. */
-  sourceObject: TaxSourceObject;
-};
-
-/** Taxable object discount. */
-export type TaxableObjectDiscount = {
-  __typename?: "TaxableObjectDiscount";
-  /** The amount of the discount. */
-  amount: Money;
-  /** The name of the discount. */
-  name?: Maybe<Scalars["String"]>;
-};
-
-export type TaxableObjectLine = {
-  __typename?: "TaxableObjectLine";
-  /** Determines if taxes are being charged for the product. */
-  chargeTaxes: Scalars["Boolean"];
-  /** The product name. */
-  productName: Scalars["String"];
-  /** The product sku. */
-  productSku?: Maybe<Scalars["String"]>;
-  /** Number of items. */
-  quantity: Scalars["Int"];
-  /** The source line related to this tax line. */
-  sourceLine: TaxSourceLine;
-  /** Price of the order line. */
-  totalPrice: Money;
-  /** Price of the single item in the order line. */
-  unitPrice: Money;
-  /** The variant name. */
-  variantName: Scalars["String"];
 };
 
 /** Represents a monetary value with taxes. In cases where taxes were not applied, net and gross values will be equal. */
@@ -21200,7 +21022,6 @@ export type WarehouseFilterInput = {
   ids?: InputMaybe<Array<Scalars["ID"]>>;
   isPrivate?: InputMaybe<Scalars["Boolean"]>;
   search?: InputMaybe<Scalars["String"]>;
-  slugs?: InputMaybe<Array<Scalars["String"]>>;
 };
 
 /**
