@@ -5,7 +5,7 @@ import { DEFAULT_LOCALE, Locale } from "@/checkout-storefront/lib/regions";
 import { reduce, snakeCase } from "lodash-es";
 import queryString from "query-string";
 import { ChangeEvent, ReactEventHandler } from "react";
-import { OperationResult } from "urql";
+import { AnyVariables, OperationResult } from "urql";
 
 export const getById =
   <T extends { id: string }>(idToCompare: string | undefined) =>
@@ -111,7 +111,7 @@ export const extractCheckoutIdFromUrl = (): string => {
   return checkoutId;
 };
 
-export const extractMutationErrors = <TData extends FormDataBase, TVars = any>(
+export const extractMutationErrors = <TData extends FormDataBase, TVars extends AnyVariables = any>(
   result: OperationResult<TData, TVars> | any // any to cover apollo client
   // mutations, to be removed once we remove apollo client from sdk
 ): [boolean, ApiErrors<TData>] => {
